fix prject open, update changelog, rename js files
